backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
